backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
supabase==1.0.4
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
httpx>=0.24.0,<0.25.0
